feat: update price fields to float and implement default database seeding for tiers and jackpots

+++ backend/app/models/jackpot.py
@@ -16,7 +16,7 @@ class Jackpot(Base):
     type = Column(String(20), nullable=False)  # midweek, mega
     dc_level = Column(Integer, nullable=False)  # 3, 4, 5, 6, 7, 10
     matches = Column(JSON, nullable=False)  # [{homeTeam, awayTeam, pick}]
-    price = Column(Integer, nullable=False)  # KES
+    price = Column(Float, nullable=False)  # KES
     regional_prices = Column(JSON, default=dict) # For dynamic overrides
 
     created_at = Column(DateTime, default=datetime.utcnow)
